Move `add_named_resource_directory` into `brioche-resources` crate
