+++ crates/brioche-packer/Cargo.toml
@@ -4,21 +4,15 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-blake3 = "1.5.1"
 brioche-autowrap = { path = "../brioche-autowrap" }
 brioche-pack = { workspace = true }
 brioche-resources = { path = "../brioche-resources" }
 bstr = "1.9.1"
 clap = { version = "4.4.11", features = ["derive"] }
 color-eyre = "0.6.3"
-copy_dir = "0.1.3"
 eyre = "0.6.12"
 globset = "0.4.14"
 goblin = "0.8.2"
-pathdiff = "0.2.1"
 runnable-core = { path = "../runnable-core" }
 serde_json = { version = "1.0.108" }
-thiserror = "1.0.51"
-tick-encoding = "0.1.2"
-ulid = "1.1.2"
 walkdir = "2.5.0"
