+++ crates/brioche-packer/src/main.rs
@@ -3,7 +3,6 @@ use std::{os::unix::fs::OpenOptionsExt as _, path::PathBuf, process::ExitCode};
 use clap::Parser;
 
 mod autowrap;
-mod resources;
 
 #[allow(clippy::large_enum_variant)]
 #[derive(Debug, Parser)]
