+++ crates/brioche-resources/Cargo.toml
@@ -6,6 +6,9 @@ edition = "2021"
 [dependencies]
 blake3 = "1.5.1"
 bstr = "1.9.1"
+copy_dir = "0.1.3"
 pathdiff = "0.2.1"
 thiserror = "1.0.61"
+tick-encoding = "0.1.2"
 ulid = "1.1.2"
+walkdir = "2.5.0"
